Abstracted paypal logic from controller. Added intial tests
